LayoutTests/imported/w3c/web-platform-tests/css/css-mixins/dashed-function-cycles-expected.txt


FAIL Local with self-cycle assert_equals: expected "PASS" but got "--f()"
FAIL Cycle reference without fallback makes result invalid assert_equals: expected "" but got "--f()"
FAIL Local with self-cycle in unused fallback assert_equals: expected "PASS" but got "--f()"
FAIL Local shadowing cyclic property --x assert_equals: expected "PASS" but got "--f()"
FAIL Local shadowing cyclic outer local --x assert_equals: expected "PASS" but got "--f()"
FAIL Argument shadowing cyclic outer local --x assert_equals: expected "10px" but got "--f()"
FAIL Arguments shadowing cyclic properties assert_equals: expected "PASS-x PASS-y" but got "--f(PASS-x, PASS-y)"
FAIL Observing property cycle locally assert_equals: expected "PASS-x PASS-y" but got "--f()"
FAIL Using cyclic values with no fallback assert_equals: expected "" but got "--f()"
FAIL Self-cycle in non-used local variable assert_equals: expected "PASS" but got "--f()"
FAIL Using cyclic value in unused fallback assert_equals: expected "PASS" but got "--f()"
FAIL Using cyclic value in unused fallback (local) assert_equals: expected "PASS" but got "--f()"
FAIL Dashed-function, self-cycle assert_equals: expected "PASS" but got "--f()"
FAIL Cycle through other function (--g) assert_equals: expected "PASS" but got "--g()"
FAIL Cycle through other function (--f) assert_equals: expected "PASS" but got "--f()"
FAIL Cycle through local, self assert_equals: expected "PASS" but got "--f()"
FAIL Cycle through unused local assert_equals: expected "PASS" but got "--f()"
FAIL Cycle through global, self assert_equals: expected "PASS" but got "--f()"
FAIL Cycle through local, other function assert_equals: expected "PASS" but got "--g()"
FAIL Cycle through local, other function, fallback in function assert_equals: expected "PASS" but got "--f()"
FAIL Cycle through various variables and other functions assert_equals: expected "PASS" but got "--g()"
FAIL Function in a cycle with its own default assert_equals: expected "PASS" but got "--f(42px)"
FAIL Cyclic defaults assert_equals: expected "42px PASS-y PASS-z" but got "--f(42px)"
FAIL Cyclic outer --b shadows custom property assert_equals: expected "PASS" but got "--f()"
FAIL Locals are function specific assert_equals: expected "10px" but got "--f()"
